Move congratulations to its own component.
