+++ settings/src/components/first-time/first-time.js
@@ -13,23 +13,9 @@ import WebAuthn from '../webauthn/webauthn';
 import BackupCodes from '../backup-codes';
 import SetupProgressBar from './setup-progress-bar';
 import Home from './home';
+import Congratulations from './congratulations';
 import WordPressLogo from './wordpress-logo';
 import { GlobalContext } from '../../script';
-
-/**
- * Check if the URL is valid. Make sure it stays on wordpress.org.
- *
- * @param  url
- * @return {boolean} Whether it's a valid URL.
- */
-const isValidUrl = ( url ) => {
-	try {
-		const { hostname } = new URL( url );
-		return hostname.endsWith( 'wordpress.org' );
-	} catch ( exception ) {
-		return false;
-	}
-};
 
 /**
  * Render the correct component based on the URL.
@@ -86,36 +72,7 @@ export default function FirstTime() {
 		},
 		congratulations: {
 			stepIndex: 3,
-			component: (
-				<div className="">
-					<h3>Two-factor authentication setup is now complete! 🎉</h3>
-					<p>
-						To ensure the highest level of security for your account, please remember to
-						keep your authentication methods up-to-date. We recommend configuring
-						multiple authentication methods to guarantee you always have access to your
-						account.
-					</p>
-					<div className="wporg-2fa__submit-actions">
-						<Button
-							onClick={ () => {
-								const redirectTo = new URLSearchParams(
-									window.location.search
-								).get( 'redirect_to' );
-
-								if ( redirectTo && isValidUrl( redirectTo ) ) {
-									window.location.href = redirectTo;
-								} else {
-									window.location.href =
-										'//profiles.wordpress.org/me/profile/edit/group/3';
-								}
-							} }
-							isPrimary
-						>
-							Continue
-						</Button>
-					</div>
-				</div>
-			),
+			component: <Congratulations />,
 		},
 	};
